Moves adaptor for mapnik and torque to renderer directories
