+++ lib/windshaft/renderers/mapnik/index.js
@@ -1,2 +1,1 @@
-module.exports.factory = require('./factory');
 module.exports.adaptor = require('./adaptor');
