+++ lib/windshaft/renderers/render_cache.js
@@ -3,7 +3,8 @@ var   _                     = require('underscore')
     , CacheEntry            = require('../cache/cache_entry')
     , Step                  = require('step')
     , tilelive              = require('tilelive')
-    , TorqueRendererFactory = require('./torque').factory;
+    , MapnikRenderer = require('./mapnik')
+    , TorqueRenderer = require('./torque');
 
 require('tilelive-mapnik').registerProtocols(tilelive);
 
@@ -39,7 +40,7 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
       bufferSize: 64
     });
 
-    var torqueFactory = new TorqueRendererFactory(); // TODO: specify options
+    var torqueFactory = new TorqueRenderer.factory(); // TODO: specify options
 
     var me = {
         serial: 0,
@@ -78,7 +79,7 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
         var poolStats = _.reduce(
             _.filter(me.renderers, function(cacheEntry) {
                 return cacheEntry.renderer
-                    && cacheEntry.renderer.constructor == TileliveAdaptor
+                    && cacheEntry.renderer.constructor == MapnikRenderer.adaptor
                     && cacheEntry.renderer.renderer._pool;
             }),
             function(_poolStats, cacheEntry) {
@@ -275,19 +276,6 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
         return opts;
     };
 
-    function TileliveAdaptor(renderer, format) {
-      this.renderer = renderer;
-      this.close = this.renderer.close.bind(this.renderer);
-      this.get = function() { return renderer; };
-      if ( format == 'png' ) {
-        this.getTile = this.renderer.getTile.bind(this.renderer);
-      }
-      else if ( format == 'grid.json' ) {
-        this.getTile = this.renderer.getGrid.bind(this.renderer);
-      }
-      else throw new Error("Unsupported format " + format);
-    }
-
     // controls the instantiation of mapnik renderer objects from mapnik XML
     me.makeRendererMapnik = function(req, callback){
         var that = this;
@@ -353,20 +341,13 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
                     req.profiler.done('tilelive-load');
                 }
                 if ( err ) throw err;
-                return new TileliveAdaptor(source, params.format);
+                return new MapnikRenderer.adaptor(source, params.format);
             },
             function returnCallback(err, source){
                 callback(err, source);
             }
         );
     };
-
-    function TorqueAdaptor(renderer) {
-      this.renderer = renderer;
-      this.close = function() { /* nothing to do */ };
-      this.get = function() { return renderer; };
-      this.getTile = this.renderer.getTile.bind(this.renderer);
-    }
 
     // controls the instantiation of mapnik renderer objects from mapnik XML
     me.makeRendererTorque = function(req, callback){
@@ -390,7 +371,7 @@ module.exports = function(options, mml_store, map_store, mapnik_opts) {
             },
             function makeAdaptor(err, renderer){
                 if ( err ) throw err;
-                return new TorqueAdaptor(renderer);
+                return new TorqueRenderer.adaptor(renderer);
             },
             function returnCallback(err, renderer){
                 callback(err, renderer);
