+++ lib/windshaft/renderers/torque/adaptor.js
@@ -1,0 +1,8 @@
+function TorqueAdaptor(renderer) {
+    this.renderer = renderer;
+    this.close = function() { /* nothing to do */ };
+    this.get = function() { return renderer; };
+    this.getTile = this.renderer.getTile.bind(this.renderer);
+}
+
+module.exports = TorqueAdaptor;
